requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-crispy-forms==1.7.2
django-filter==2.0.0
django-import-export==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jsonapi==2.6.0
et-xmlfile==1.0.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.13
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==2.5.12
Pillow==5.3.0
protobuf==3.6.1
pytz==2018.7
PyYAML==3.13
requests==2.21.0
simplejson==3.16.0
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
